pyproject.toml
[tool.poetry]
name = "caldal-backend"
version = "0.1.0"
description = ""
authors = ["Seung-Yoon Oh <feynubrick@proton.me>"]
readme = "README.md"
package-mode = false


[tool.poetry.dependencies]
python = "^3.12"
django-ninja = "^1.3.0"
python-dotenv = "^1.0.1"
pre-commit = "^3.8.0"
pyhumps = "^3.8.0"
orjson = "^3.10.7"
django-ninja-jwt = "^5.3.3"
django-ninja-extra = "^0.21.4"
google-api-python-client = "^2.149.0"
google-auth = "^2.35.0"
google-auth-oauthlib = "^1.2.1"
google-auth-httplib2 = "^0.2.0"
requests = "^2.32.3"
django-extensions = "^3.2.3"
psycopg = "^3.2.3"
gunicorn = "^23.0.0"
uvicorn = "^0.32.0"
pytest-bdd = "^7.3.0"
pytest-django = "^4.9.0"
email-validator = "^2.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
